+++ src/main/java/edu/clemson/cs/rsrg/vcgeneration/utilities/Utilities.java
@@ -49,6 +49,7 @@ import edu.clemson.cs.rsrg.treewalk.TreeWalkerVisitor;
 import edu.clemson.cs.rsrg.typeandpopulate.typereasoning.TypeGraph;
 import edu.clemson.cs.rsrg.vcgeneration.VCGenerator;
 import edu.clemson.cs.rsrg.vcgeneration.sequents.Sequent;
+import edu.clemson.cs.rsrg.vcgeneration.utilities.formaltoactual.InstantiatedFacilityDecl;
 import java.util.*;
 import org.jgrapht.DirectedGraph;
 import org.jgrapht.alg.interfaces.ShortestPathAlgorithm;
