+++ src/main/java/edu/clemson/cs/rsrg/vcgeneration/utilities/formaltoactual/EnhancementSpecRealizItemMap.java
@@ -10,7 +10,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.rsrg.vcgeneration.utilities;
+package edu.clemson.cs.rsrg.vcgeneration.utilities.formaltoactual;
 
 import edu.clemson.cs.rsrg.absyn.declarations.facilitydecl.FacilityDec;
 import edu.clemson.cs.rsrg.absyn.expressions.Exp;
